Moving the reference processing code from the client project to a separate class library.
