+++ WikipediaReferenceCleaner.sln
@@ -1,11 +1,13 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
-VisualStudioVersion = 17.12.35527.113 d17.12
+VisualStudioVersion = 17.12.35527.113
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WikipediaReferenceCleaner", "WikipediaReferenceCleaner\WikipediaReferenceCleaner.csproj", "{7FECCABA-35C4-4221-9F7A-AD1C59807DAD}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WikipediaReferenceCleanerTests", "WikipediaReferenceCleanerTests\WikipediaReferenceCleanerTests.csproj", "{EC4B7E6C-2A6F-4561-B720-0F549F4BD7EF}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ReferenceProcessing", "ReferenceProcessing\ReferenceProcessing.csproj", "{97B77C91-D06B-4843-A4B1-F37353FD1931}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -21,6 +23,10 @@ Global
 		{EC4B7E6C-2A6F-4561-B720-0F549F4BD7EF}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{EC4B7E6C-2A6F-4561-B720-0F549F4BD7EF}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{EC4B7E6C-2A6F-4561-B720-0F549F4BD7EF}.Release|Any CPU.Build.0 = Release|Any CPU
+		{97B77C91-D06B-4843-A4B1-F37353FD1931}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{97B77C91-D06B-4843-A4B1-F37353FD1931}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{97B77C91-D06B-4843-A4B1-F37353FD1931}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{97B77C91-D06B-4843-A4B1-F37353FD1931}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
